+++ src/filestove/stove.hpp
@@ -25,7 +25,7 @@ public:
     bool cook();
 
     void resetReadCount() noexcept;
-    std::size_t readCount() const noexcept;
+    std::uintmax_t readCount() const noexcept;
 
     bool isDone() const noexcept;
 };
